Dropping images: Extract implementation to separate file
https://github.com/nteract/nteract/issues/5081
